split out making gif
